--- .github/workflows/command_line_taxvamb.yml
@@ -1,50 +1,0 @@
-# This workflow will install Python dependencies, run tests and lint with a single version of Python
-# For more information see: https://docs.github.com/en/actions/automating-builds-and-tests/building-and-testing-python
-
-name: Command line interface tests
-
-on:
-  push:
-    branches: [ "master" ]
-  pull_request:
-    branches: [ "master" ]
-  workflow_dispatch:
-    inputs:
-
-permissions:
-  contents: read
-
-jobs:
-  build:
-    runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        python-version: ["3.11"]
-
-    steps:
-    - uses: actions/checkout@v3
-    - name: Set up Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v3
-      with:
-        python-version: ${{ matrix.python-version }}
-        cache: 'pip' # caching pip dependencies
-        cache-dependency-path: '**/pyproject.toml'
-    - name: Download fixtures
-      run: |
-        wget https://www.dropbox.com/scl/fi/ivsqbao9we1lnyyyh43cc/ci_data.zip\?rlkey\=b9lcqxs63iv2409q5a7q1oh9z\&dl\=0 -O ci_data.zip
-        unzip ci_data.zip
-    - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip
-        pip install flake8 pytest
-        pip install -e .
-    - name: Run VAMB
-      run: |
-        vamb --outdir outdir_vamb --fasta catalogue_mock.fna.gz --rpkm abundance_mock.npz -l 32 -e 10 -q 2 -o C --minfasta 200000
-        ls -la outdir_vamb
-        cat outdir_vamb/log.txt
-    - name: Run TaxVAMB
-      run: |
-        vamb --model vaevae --outdir outdir_taxvamb --fasta catalogue_mock.fna.gz --rpkm abundance_mock.npz --taxonomy taxonomy_mock.tsv -pe 10 -e 10 -q 2 -o C --minfasta 200000
-        ls -la outdir_taxvamb
-        cat outdir_taxvamb/log.txt
